page profile add and work

+++ handlers/start_server.go
@@ -22,7 +22,7 @@ var tmpl_login *template.Template
 var tmpl_main_page *template.Template
 var tmpl_admin_pannel *template.Template
 var tmpl_create_poste *template.Template
-var tmpl_create_image *template.Template
+var tmpl_profile *template.Template
 var db *sql.DB
 var sessions = map[string]string{}
 var sessionsMutex sync.Mutex
@@ -80,6 +80,11 @@ func StartServer() {
 	}
 
 	tmpl_create_poste, err = template.New("createPost").ParseFiles(filepath.Join(wd, "Static", "Templates", "createPost.html"))
+	if err != nil {
+		panic(err)
+	}
+
+	tmpl_profile, err = template.New("profile").ParseFiles(filepath.Join(wd, "Static", "Templates", "profile.html"))
 	if err != nil {
 		panic(err)
 	}
@@ -170,6 +175,15 @@ func StartServer() {
 		}
 	})
 
+	http.HandleFunc("/profile", func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/profile" {
+			data := handleProfile(w, r)
+			tmpl_profile.Execute(w, data)
+		} else {
+			fileServer.ServeHTTP(w, r)
+		}
+	})
+
 	http.HandleFunc("/registerUser", handleRegister)
 	http.HandleFunc("/loginUser", handleLogin)
 	http.HandleFunc("/upgradeRank", handleRankUp)
@@ -184,6 +198,18 @@ func StartServer() {
 }
 
 func createTables(db *sql.DB) {
+	/*dropPostTableSQL := `DROP TABLE IF EXISTS Post;`
+	_, err := db.Exec(dropPostTableSQL)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	dropLikeTableSQL := `DROP TABLE IF EXISTS PostLikes;`
+	_, err = db.Exec(dropLikeTableSQL)
+	if err != nil {
+		log.Fatal(err)
+	}*/
+
 	createAccountTableSQL := `CREATE TABLE IF NOT EXISTS Account (
         id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
         username TEXT NOT NULL,
@@ -388,6 +414,129 @@ func handleHome(w http.ResponseWriter, r *http.Request) map[string]interface{} {
 	return data
 }
 
+func handleProfile(w http.ResponseWriter, r *http.Request) map[string]interface{} {
+	username := getSessionUsername(r)
+	userID := getUserIDByUsername(username)
+
+	// Requête SQL pour récupérer les posts aimés par l'utilisateur
+	query := `
+        SELECT 
+            p.id, 
+            p.post_name, 
+            p.creator_id, 
+            p.post_message, 
+            p.category_name, 
+            p.likes, 
+            p.dislikes
+        FROM 
+            Post p
+        LEFT JOIN 
+            PostLikes pl ON p.id = pl.post_id
+        WHERE 
+            pl.user_id = ? AND pl.liked = 1
+        ORDER BY 
+            p.id DESC`
+
+	rows, err := db.Query(query, userID)
+	if err != nil {
+		log.Printf("Error fetching posts: %v", err)
+		http.Error(w, "Error fetching posts", http.StatusInternalServerError)
+		return nil
+	}
+	defer rows.Close()
+
+	var postsLike []map[string]interface{}
+	for rows.Next() {
+		var id, likes, dislikes int
+		var postName, postMessage, categoryName string
+		var creatorID int // creatorID should be an int as it references Account(id)
+		err := rows.Scan(&id, &postName, &creatorID, &postMessage, &categoryName, &likes, &dislikes)
+		if err != nil {
+			log.Printf("Error scanning post: %v", err)
+			http.Error(w, "Error scanning post", http.StatusInternalServerError)
+			return nil
+		}
+
+		post := map[string]interface{}{
+			"ID":           id,
+			"PostName":     postName,
+			"CreatorID":    creatorID,
+			"PostMessage":  postMessage,
+			"CategoryName": categoryName,
+			"Likes":        likes,
+			"Dislikes":     dislikes,
+		}
+		postsLike = append(postsLike, post)
+	}
+
+	// Récupérer les postes de la base de données
+	rows, err = db.Query(`SELECT 
+            p.id, 
+            p.post_name, 
+            p.creator_id, 
+            p.post_message, 
+            p.category_name, 
+            IFNULL(likeCount, 0) as likeCount, 
+            IFNULL(dislikeCount, 0) as dislikeCount
+        FROM 
+            Post p 
+        LEFT JOIN (
+            SELECT 
+                post_id, 
+                SUM(CASE WHEN liked = 1 THEN 1 ELSE 0 END) as likeCount, 
+                SUM(CASE WHEN liked = 0 THEN 1 ELSE 0 END) as dislikeCount 
+            FROM 
+                PostLikes 
+            GROUP BY 
+                post_id
+        ) pl 
+        ON p.id = pl.post_id 
+        ORDER BY p.id DESC`)
+	if err != nil {
+		http.Error(w, "Error fetching posts", http.StatusInternalServerError)
+		return nil
+	}
+	defer rows.Close()
+	isEqID := false
+
+	var posts []map[string]interface{}
+	for rows.Next() {
+		var id, likes, dislikes, creatorID int
+		var postName, postMessage, category_name string
+		err := rows.Scan(&id, &postName, &creatorID, &postMessage, &category_name, &likes, &dislikes)
+		if err != nil {
+			http.Error(w, "Error scanning post", http.StatusInternalServerError)
+			return nil
+		}
+
+		if userID == creatorID {
+			isEqID = true
+		} else {
+			isEqID = false
+		}
+		post := map[string]interface{}{
+			"ID":           id,
+			"PostName":     postName,
+			"CreatorID":    creatorID,
+			"PostMessage":  postMessage,
+			"categoryName": category_name,
+			"LikeCount":    likes,
+			"DislikeCount": dislikes,
+			"IsEqID":       isEqID,
+		}
+		posts = append(posts, post)
+	}
+
+	data := map[string]interface{}{
+		"Username":  username,
+		"Posts":     posts,
+		"PostsLike": postsLike,
+		"UserId":    userID,
+	}
+
+	return data
+}
+
 func handleLikePost(w http.ResponseWriter, r *http.Request) {
 	userID := getSessionUserID(r)
 	if userID == 0 {
@@ -407,7 +556,8 @@ func handleLikePost(w http.ResponseWriter, r *http.Request) {
 
 	// Check if the user has already liked/disliked the post
 	var existingLikeID int
-	err := db.QueryRow(`SELECT id FROM PostLikes WHERE post_id = ? AND user_id = ?`, postID, userID).Scan(&existingLikeID)
+	var existingLiked bool
+	err := db.QueryRow(`SELECT id, liked FROM PostLikes WHERE post_id = ? AND user_id = ?`, postID, userID).Scan(&existingLikeID, &existingLiked)
 
 	if err == sql.ErrNoRows {
 		// No existing like/dislike, insert a new one
@@ -417,16 +567,28 @@ func handleLikePost(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 	} else if err == nil {
-		// Existing like/dislike found, remove it
-		_, err = db.Exec(`DELETE FROM PostLikes WHERE id = ?`, existingLikeID)
-		if err != nil {
-			http.Error(w, "Error unliking post", http.StatusInternalServerError)
-			return
+		if existingLiked == liked {
+			// Existing like/dislike found and matches the current action, remove it (toggle off)
+			_, err = db.Exec(`DELETE FROM PostLikes WHERE id = ?`, existingLikeID)
+			if err != nil {
+				http.Error(w, "Error unliking post", http.StatusInternalServerError)
+				return
+			}
+		} else {
+			// Existing like/dislike found but does not match the current action, update it (toggle switch)
+			_, err = db.Exec(`UPDATE PostLikes SET liked = ? WHERE id = ?`, liked, existingLikeID)
+			if err != nil {
+				http.Error(w, "Error updating like status", http.StatusInternalServerError)
+				return
+			}
 		}
 	} else {
 		http.Error(w, "Error checking like status", http.StatusInternalServerError)
 		return
 	}
+
+	// Update the likes and dislikes in the Post table
+	updatePostLikes(postID)
 
 	// Return the updated like/dislike count for the post
 	var likeCount, dislikeCount int
@@ -446,6 +608,18 @@ func handleLikePost(w http.ResponseWriter, r *http.Request) {
 	jsonResponse, _ := json.Marshal(response)
 	w.Header().Set("Content-Type", "application/json")
 	w.Write(jsonResponse)
+}
+
+func updatePostLikes(postID string) {
+	_, err := db.Exec(`
+        UPDATE Post
+        SET likes = (SELECT COUNT(*) FROM PostLikes WHERE post_id = ? AND liked = 1),
+            dislikes = (SELECT COUNT(*) FROM PostLikes WHERE post_id = ? AND liked = 0)
+        WHERE id = ?
+    `, postID, postID, postID)
+	if err != nil {
+		log.Printf("Error updating post likes: %v", err)
+	}
 }
 
 func getSessionUserID(r *http.Request) int {
@@ -546,8 +720,10 @@ func createPost(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	username := getSessionUsername(r)
+
 	postName := r.FormValue("postName")
-	creatorID := getSessionUsername(r)
+	creatorID := getUserIDByUsername(username)
 	postMessage := r.FormValue("postMessage")
 	category_name := r.FormValue("category_name")
 
@@ -684,4 +860,14 @@ func getSessionRank(r *http.Request) (string, error) {
 	}
 
 	return rank, nil
+}
+
+func getUserIDByUsername(username string) int {
+	var userID int
+	err := db.QueryRow("SELECT id FROM Account WHERE username = ?", username).Scan(&userID)
+	if err != nil {
+		log.Println("Error fetching user ID:", err)
+		return 0 // Handle this case appropriately
+	}
+	return userID
 }
